Replace ChangeColorMessage definition with actual RGB-compatible types

+++ listings/ch06-enums-and-pattern-matching/no-listing-04-structs-similar-to-message-enum/src/main.rs
@@ -5,7 +5,7 @@ struct MoveMessage {
     y: i32,
 }
 struct WriteMessage(String); // tuple struct
-struct ChangeColorMessage(i32, i32, i32); // tuple struct
+struct ChangeColorMessage(u8, u8, u8); // tuple struct
                                           // ANCHOR_END: here
 
 fn main() {}
